small normie stuff cuz i need space
